# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/translations/scanner/sync": {
      "post": {
        "operationId": "TranslationScannerController_sync",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncTranslationScanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Translation scan synchronized successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Synchronize a Shopify translation scan",
        "tags": [
          "Translation Scanner"
        ]
      }
    },
    "/api/v1/translations/scanner/summary": {
      "get": {
        "operationId": "TranslationScannerController_summary",
        "parameters": [
          {
            "name": "resourceType",
            "required": true,
            "in": "query",
            "schema": {
              "example": "PRODUCT",
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "schema": {
              "example": "it",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Translation scan summary returned successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get translation scan summary for one resource type",
        "tags": [
          "Translation Scanner"
        ]
      }
    },
    "/api/v1/translations/scanner/overview": {
      "get": {
        "description": "Returns the translation status grouped into content, custom data, commerce, theme and system resources.",
        "operationId": "TranslationScannerController_overview",
        "parameters": [
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "schema": {
              "example": "it",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Global translation scanner overview returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationScannerOverviewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the global translation scanner overview",
        "tags": [
          "Translation Scanner"
        ]
      }
    },
    "/api/v1/translations/scanner/resource-types": {
      "get": {
        "description": "Returns scanner statistics for every supported Shopify translatable resource type, including resource types with zero content.",
        "operationId": "TranslationScannerController_resourceTypesOverview",
        "parameters": [
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "schema": {
              "example": "it",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resource type translation statistics returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationResourceTypesOverviewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get translation scanner statistics by Shopify resource type",
        "tags": [
          "Translation Scanner"
        ]
      }
    },
    "/api/v1/translations/resources": {
      "get": {
        "operationId": "TranslationScannerController_resources",
        "parameters": [
          {
            "name": "resourceType",
            "required": true,
            "in": "query",
            "schema": {
              "example": "PRODUCT",
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "schema": {
              "example": "it",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scanned translation resources returned successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List scanned translation resources",
        "tags": [
          "Translation Scanner"
        ]
      }
    },
...
  },
  "info": {
    "title": "ACA Locale API",
    "description": "Technical API documentation for ACA Locale.\n\nACA Locale is an AI-powered localization management platform.\n\nThis specification is generated automatically from the NestJS backend.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TranslationSourceDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "title"
          },
          "type": {
            "type": "string",
            "example": "SINGLE_LINE_TEXT_FIELD"
          },
          "locale": {
            "type": "string",
            "example": "en"
          },
          "value": {
            "type": "string",
            "example": "Gift Card"
          },
          "digest": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "type",
          "locale",
          "value",
          "digest"
        ]
      },
      "ExistingTranslationDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "title"
          },
          "locale": {
            "type": "string",
            "example": "it"
          },
          "value": {
            "type": "string"
          },
          "outdated": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string"
          },
          "market": {
            "type": "object"
          }
        },
        "required": [
          "key",
          "locale",
          "value",
          "outdated"
        ]
      },
      "TranslationResourceScanDto": {
        "type": "object",
        "properties": {
          "resourceId": {
            "type": "string",
            "example": "gid://shopify/Product/8882758975625"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranslationSourceDto"
            }
          },
          "translations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExistingTranslationDto"
            }
          }
        },
        "required": [
          "resourceId",
          "content",
          "translations"
        ]
      },
      "SyncTranslationScanDto": {
        "type": "object",
        "properties": {
          "resourceType": {
            "type": "string",
            "example": "PRODUCT"
          },
          "targetLocale": {
            "type": "string",
            "example": "it"
          },
          "resources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranslationResourceScanDto"
            }
          },
          "completeScan": {
            "type": "boolean",
            "description": "Indicates that the payload contains the complete Shopify result set for this resource type. When true, stale local resources are removed.",
            "example": true,
            "default": false
          }
        },
        "required": [
          "resourceType",
          "targetLocale",
          "resources"
        ]
      },
      "TranslationOverallSummaryDto": {
        "type": "object",
        "properties": {
          "resourceTypes": {
            "type": "number",
            "example": 30
          },
          "resourceTypesWithContent": {
            "type": "number",
            "example": 22
          },
          "resources": {
            "type": "number",
            "example": 126
          },
          "fields": {
            "type": "number",
            "example": 9173
          },
          "actionableFields": {
            "type": "number",
            "example": 9163
          },
          "emptySource": {
            "type": "number",
            "example": 10
          },
          "missing": {
            "type": "number",
            "example": 9163
          },
          "translated": {
            "type": "number",
            "example": 0
          },
          "outdated": {
            "type": "number",
            "example": 0
          },
          "coverage": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "resourceTypes",
          "resourceTypesWithContent",
          "resources",
          "fields",
          "actionableFields",
          "emptySource",
          "missing",
          "translated",
          "outdated",
          "coverage"
        ]
      },
      "TranslationGroupSummaryDto": {
        "type": "object",
        "properties": {
          "group": {
            "type": "string",
            "enum": [
              "CONTENT",
              "CUSTOM_DATA",
              "COMMERCE",
              "THEME",
              "SYSTEM",
              "OTHER"
            ],
            "example": "CONTENT"
          },
          "resourceTypes": {
            "type": "number",
            "example": 12
          },
          "resourceTypesWithContent": {
            "type": "number",
            "example": 8
          },
          "resources": {
            "type": "number",
            "example": 96
          },
          "fields": {
            "type": "number",
            "example": 145
          },
          "actionableFields": {
            "type": "number",
            "example": 143
          },
          "emptySource": {
            "type": "number",
            "example": 2
          },
          "missing": {
            "type": "number",
            "example": 143
          },
          "translated": {
            "type": "number",
            "example": 0
          },
          "outdated": {
            "type": "number",
            "example": 0
          },
          "coverage": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "group",
          "resourceTypes",
          "resourceTypesWithContent",
          "resources",
          "fields",
          "actionableFields",
          "emptySource",
          "missing",
          "translated",
          "outdated",
          "coverage"
        ]
      },
      "TranslationScannerOverviewDto": {
        "type": "object",
        "properties": {
          "shopifyDomain": {
            "type": "string",
            "example": "aca-locale-dev-cfoxunuo.myshopify.com"
          },
          "targetLocale": {
            "type": "string",
            "example": "it"
          },
          "summary": {
            "$ref": "#/components/schemas/TranslationOverallSummaryDto"
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranslationGroupSummaryDto"
            }
          }
        },
        "required": [
          "shopifyDomain",
          "targetLocale",
          "summary",
          "groups"
        ]
      },
      "TranslationResourceTypeSummaryDto": {
        "type": "object",
        "properties": {
          "resourceType": {
            "type": "string",
            "example": "PRODUCT"
          },
          "group": {
            "type": "string",
            "enum": [
              "CONTENT",
              "CUSTOM_DATA",
              "COMMERCE",
              "THEME",
              "SYSTEM",
              "OTHER"
            ],
            "example": "CONTENT"
          },
          "resources": {
            "type": "number",
            "example": 17
          },
          "fields": {
            "type": "number",
            "example": 57
          },
          "actionableFields": {
            "type": "number",
            "example": 56
          },
          "emptySource": {
            "type": "number",
            "example": 1
          },
          "missing": {
            "type": "number",
            "example": 56
          },
          "translated": {
            "type": "number",
            "example": 0
          },
          "outdated": {
            "type": "number",
            "example": 0
          },
          "coverage": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "resourceType",
          "group",
          "resources",
          "fields",
          "actionableFields",
          "emptySource",
          "missing",
          "translated",
          "outdated",
          "coverage"
        ]
      },
      "TranslationResourceTypesOverviewDto": {
        "type": "object",
        "properties": {
          "shopifyDomain": {
            "type": "string",
            "example": "aca-locale-dev-cfoxunuo.myshopify.com"
          },
          "targetLocale": {
            "type": "string",
            "example": "it"
          },
          "resourceTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranslationResourceTypeSummaryDto"
            }
          }
        },
        "required": [
          "shopifyDomain",
          "targetLocale",
          "resourceTypes"
        ]
      },
...
    }
  }
}